+++ R/tsmethods-package.R
@@ -1,5 +1,6 @@
 #' @keywords internal
 #' @import methods
+#' @import data.table
 #' @importFrom stats na.omit ts median quantile fitted
 #' @importFrom utils tail head data
 #' @importFrom graphics hist grid legend lines par plot points abline axis axis.Date axis.POSIXct box polygon layout